+++ _engine/code/include/global/core/data/input/ustring.hpp
@@ -40,17 +40,17 @@ namespace engine
 				this->pos = data.len() - 1;
 			}
 
-			void skip(int32_t pos) final
+			void skip(int64_t pos) final
 			{
-				this->pos = std::max(0, std::min(static_cast<int32_t>(data.len()) - 1, this->pos + pos));
+				this->pos = std::max(0ll, std::min(static_cast<int64_t>(data.len()) - 1, static_cast<int64_t>(this->pos) + pos));
 			}
-			void go_back(int32_t pos) final
+			void go_back(int64_t pos) final
 			{
-				this->pos = std::max(0, std::min(static_cast<int32_t>(data.len()) - 1, this->pos - pos));
+				this->pos = std::max(0ll, std::min(static_cast<int64_t>(data.len()) - 1, static_cast<int64_t>(this->pos) - pos));
 			}
-			uint32_t read(uint8_t * buffer, uint32_t size) final
+			uint64_t read(uint8_t * buffer, uint64_t size) final
 			{
-				uint32_t len = std::min(data.len(), this->pos + size) - this->pos;
+				uint64_t len = std::min(data.len(), static_cast<int64_t>(this->pos) + size) - this->pos;
 				memcpy(buffer, data.get_cstring() + this->pos, len);
 				pos += len;
 				return len;
@@ -60,14 +60,14 @@ namespace engine
 				return pos >= data.len();
 			}
 
-			uint32_t position() final
+			uint64_t position() final
 			{
 				return pos;
 			}
 
 		private:
 
-			int32_t pos;
+			int64_t pos;
 
 			ustring_t data;
 		};
