+++ _engine/code/include/global/core/data/output/file.hpp
@@ -30,7 +30,7 @@ namespace engine
 
 			}
 
-			uint32_t write(const uint8_t * buffer, uint32_t size) final
+			uint64_t write(const uint8_t * buffer, uint64_t size) final
 			{
 				return output->write(buffer, size);
 			}
